refactor: decompose league page + code review fixes
- Extract lib/format.ts (formatting helpers)
- Extract lib/league-utils.ts (sort/filter/presets config)
- Extract components/league/ (cells, card-thumb, table-skeleton, sort-indicator)
- Lift SortableHead/PlainHead out of render scope (prevent remount)
- Fix NaN sort comparator for zero-cost cards
- Fix stale profitRange on league navigation
- Replace Math.min/max spread with reduce (stack overflow safety)
- Combine allProfits/min/max into single-pass memo
- Simplify ProfitCell prefix logic
- Remove dead profitMax < Infinity guard
- Update Kody workflow to target master + develop branches
